+++ NEWS.md
@@ -1,3 +1,5 @@
+# glydet (development version)
+
 # glydet 0.11.1
 
 ## Minor improvements and bug fixes