+++ NEWS.md
@@ -1,3 +1,5 @@
+# glydet (development version)
+
 # glydet 0.12.0
 
 ## Breaking changes